requirements.txt
discord-py-interactions
aiohappyeyeballs==2.4.4
aiohttp==3.11.9
aiosignal==1.3.1
attrs==24.2.0
certifi==2024.8.30
charset-normalizer==3.4.0
croniter==5.0.1
discord-typings==0.9.0
emoji==2.14.0
frozenlist==1.5.0
h11==0.14.0
idna==3.10
multidict==6.1.0
outcome==1.3.0.post0
propcache==0.2.1
PySocks==1.7.1
python-dateutil==2.9.0.post0
pytz==2024.2
requests==2.32.3
selenium==4.27.1
setuptools==75.6.0
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
tomli==2.2.1
trio==0.27.0
trio-websocket==0.11.1
typing_extensions==4.12.2
urllib3==2.2.3
websocket-client==1.8.0
websockets==14.1
wsproto==1.2.0
yarl==1.18.3
moviepy==2.1.1
webdriver-manager
yt-dlp
boto3
botocore